pyproject.toml
[project]
name = "atlan-oracle-app"
version = "0.1.0"
description = "Oracle Enterprise Connector: Built using Atlan Application SDK"
authors = [{ name = "Your Name", email = "your.email@company.com" }]
requires-python = ">=3.11"
license = "Apache-2.0"
readme = "README.md"
dependencies = [
    "atlan-application-sdk[daft,iam-auth,sqlalchemy,tests,workflows,pandas]==0.1.1rc10",
    "oracledb>=1.4.0",
    "jinja2>=3.1.2",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.8.0",
    "isort>=5.13.2",
    "watchdog>=6.0.0",
    "requests",
    "poethepoet>=0.34.0",
]
test = [
    "pytest>=8.3.3",
    "coverage>=7.6.1",
    "scalene>=1.5.20",
    "hypothesis>=6.98.8",
    "pytest-asyncio>=0.24.0"
]

[tool.poe.tasks]
download-components = "python scripts/download_components.py --ref v0.1.1rc10"

# Dapr and Temporal service tasks
start-dapr = "dapr run --enable-api-logging --log-level debug --app-id app --app-port 3000 --dapr-http-port 3500 --dapr-grpc-port 50001 --dapr-http-max-request-size 1024 --resources-path components"
start-temporal = "temporal server start-dev --db-filename /tmp/temporal.db"
start-deps.shell = "poe start-dapr & poe start-temporal &"
stop-deps.shell = "lsof -ti:3000,3500,7233,50001 | xargs kill -9 2>/dev/null || true"

[tool.hatch.build.targets.sdist]
include = ["./**/*.py"]

[tool.hatch.build.targets.wheel]
include = ["./**/*.py"]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
venvPath = "."
venv = ".venv"
typeCheckingMode = "standard"
exclude = [".github/**"]
reportMissingTypeStubs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.coverage.run]
omit = ["tests/**"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]
fail_under = 60
